+++ packages/sdk/index.d.ts
@@ -2,11 +2,12 @@ import Sdk from './dist/types/src/sdk';
 
 export const globalConfig: Sdk['globalConfig'];
 export const base: Sdk['base'];
-export const session: Sdk['session'];
 export const installationId: Sdk['installationId'];
 export const viewport: Sdk['viewport'];
 export const runInfo: Sdk['runInfo'];
-export const cursor: Sdk['cursor'];
 export const settingsButton: Sdk['settingsButton'];
 export const undoRedo: Sdk['undoRedo'];
 export const reload: Sdk['reload'];
+
+export const cursor: Sdk['cursor'];
+export const session: Sdk['session'];