+++ packages/sdk/test/ui/initialize_block.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import MockAirtableInterface from '../airtable_interface_mocks/mock_airtable_interface_internal';
 import {initializeBlock} from '../../src/ui/ui';
+import {__resetHasBeenInitialized} from '../../src/ui/initialize_block';
 
 let mockAirtableInterface: jest.Mocked<MockAirtableInterface>;
 jest.mock('../../src/injected/airtable_interface', () => ({
@@ -14,7 +15,13 @@ jest.mock('../../src/injected/airtable_interface', () => ({
 }));
 
 describe('initializeBlock', () => {
+    afterEach(() => {
+        __resetHasBeenInitialized();
+    });
     it("functions without explicitly importing the SDK's main entry point", () => {
         initializeBlock(() => <div />);
+    });
+    it('functions as a view and a block through one entry point', () => {
+        initializeBlock({dashboard: () => <div />, view: () => <div />});
     });
 });