+++ packages/sdk/CHANGELOG.md
@@ -9,9 +9,16 @@ Not every commit needs to result in a change to this file (e.g. docs and chore c
 commit that affects the code in a way that consumers might care about should include edits to the
 'Unreleased' section though. Breaking changes should be prefixed with `**BREAKING:**`.
 
-## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@1.3.0...HEAD)
+## [Unreleased](https://github.com/airtable/blocks/compare/@airtable/blocks@1.4.0...HEAD)
 
 No changes.
+
+## [1.4.0](https://github.com/airtable/blocks/compare/@airtable/blocks@1.3.0...@airtable/blocks@1.4.0) - 2021-01-19
+
+-   **DEPRECATED:** importing the Base's Cursor instance from the main entrypoint, e.g.
+    `import {cursor} from '@airtable/blocks';`. Use the `useCursor` React Hook instead.
+-   **DEPRECATED:** importing the Base's Session instance from the main entrypoint, e.g.
+    `import {session} from '@airtable/blocks';`. Use the `useSession` React Hook instead.
 
 ## [1.3.0](https://github.com/airtable/blocks/compare/@airtable/blocks@1.2.5...@airtable/blocks@1.3.0) - 2021-01-07